+++ Demo/VComponentsDemo/Scenes/Demos/Containers/VDisclosureGroupDemoView.swift
@@ -49,19 +49,14 @@ struct VDisclosureGroupDemoView: View {
                 isExpanded: $isExpanded,
                 headerTitle: "Lorem Ipsum",
                 content: {
-                    VStaticList(
-                        uiModel: {
-                            var uiModel: VStaticListUIModel = .init()
-                            uiModel.layout.showsFirstSeparator = false
-                            uiModel.layout.showsLastSeparator = false
-                            return uiModel
-                        }(),
-                        data: 0..<10,
-                        content: { num in
-                            Text(String(num))
-                                .frame(maxWidth: .infinity, alignment: .leading)
-                        }
-                    )
+                    LazyVStack(spacing: 0, content: {
+                        ForEach(0..<10, content: { num in
+                            VListRow(separator: .noFirstAndLastSeparators(isFirst: num == 0), content: {
+                                Text(String(num))
+                                    .frame(maxWidth: .infinity, alignment: .leading)
+                            })
+                        })
+                    })
                 }
             )
                 .disabled(!isEnabled)