+++ Demo/VComponentsDemo/Scenes/Demos/Modals/VModalDemoView.swift
@@ -105,19 +105,15 @@ struct VModalDemoView: View {
     
     private var modalContent: some View {
         ZStack(content: {
-            VList(
-                uiModel: {
-                    var uiModel: VListUIModel = .init()
-                    uiModel.layout.showsFirstSeparator = false
-                    uiModel.layout.showsLastSeparator = false
-                    return uiModel
-                }(),
-                data: 0..<20,
-                content: { num in
-                    Text(String(num))
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                }
-            )
+            List(content: {
+                ForEach(0..<20, content: { num in
+                    VListRow(separator: .noFirstAndLastSeparators(isFirst: num == 0), content: {
+                        Text(String(num))
+                            .frame(maxWidth: .infinity, alignment: .leading)
+                    })
+                })
+            })
+                .vListStyle()
 
             if dismissType.isEmpty {
                 NoDismissTypeWarningView(onDismiss: { isPresented = false })