+++ Demo/VComponentsDemo/Scenes/Demos/Modals/VSideBarDemoView.swift
@@ -88,21 +88,17 @@ struct VSideBarDemoView: View {
 
     private var sideBarContent: some View {
         ZStack(content: {
-            VList(
-                uiModel: {
-                    var uiModel: VListUIModel = .init()
-                    uiModel.layout.showsFirstSeparator = false
-                    uiModel.layout.showsLastSeparator = false
-                    return uiModel
-                }(),
-                data: 0..<40,
-                content: { num in
-                    Text(String(num))
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                }
-            )
+            List(content: {
+                ForEach(0..<40, content: { num in
+                    VListRow(separator: .noFirstAndLastSeparators(isFirst: num == 0), content: {
+                        Text(String(num))
+                            .frame(maxWidth: .infinity, alignment: .leading)
+                    })
+                })
+            })
+                .vListStyle()
                 .padding(.vertical, 1) // Fixes SwiftUI `ScrollView` safe area bug
-            
+
             if dismissType.isEmpty {
                 NoDismissTypeWarningView(onDismiss: { isPresented = false })
             }