+++ Sources/VComponents/Components/Modals/Bottom Sheet/VBottomSheet.swift
@@ -362,19 +362,15 @@ struct VBottomSheet_Previews: PreviewProvider {
                 isPresented: $isPresented,
                 headerTitle: "Lorem Ipsum Dolor Sit Amet",
                 content: {
-                    VList(
-                        uiModel: {
-                            var uiModel: VListUIModel = .init()
-                            uiModel.layout.showsFirstSeparator = false
-                            uiModel.layout.showsLastSeparator = false
-                            return uiModel
-                        }(),
-                        data: 0..<20,
-                        content: { num in
-                            Text(String(num))
-                                .frame(maxWidth: .infinity, alignment: .leading)
-                        }
-                    )
+                    List(content: {
+                        ForEach(0..<20, content: { num in
+                            VListRow(separator: .noFirstAndLastSeparators(isFirst: num == 0), content: {
+                                Text(String(num))
+                                    .frame(maxWidth: .infinity, alignment: .leading)
+                            })
+                        })
+                    })
+                        .vListStyle()
                 }
             )
     }